+++ application/controllers/admin/Staff.php
@@ -134,7 +134,9 @@ class staff extends MY_Controller
 		$status = $this->input->post('status');
 
 		if ($foto != "") {
-			unlink("./upload/staff/" . $old_foto);
+			if ($old_foto != "image.jpg") {
+				unlink("./upload/staff/" . $old_foto);
+			}
 			//upload foto
 			define('MB', 1048576);
 			if ($_FILES['foto']['size'] > 5 * MB) { // JIKA FILE DI UPLOAD OLEH USER
